move schema registry client impl from impl package and make constructor package private
